+++ backend/src/xfd_django/xfd_api/api_methods/notification.py
@@ -8,9 +8,14 @@ from fastapi import HTTPException, status
 from xfd_api.models import Notification
 from xfd_api.schema_models.notification import Notification as NotificationSchema
 
+from ..auth import get_org_memberships, is_global_view_admin
+
 
 def post(data, current_user):
     """POST LOGIC"""
+    # Check if user is GlobalViewAdmin or has memberships
+    if not is_global_view_admin(current_user) and not get_org_memberships(current_user):
+        raise HTTPException(status_code=403, detail="Unauthorized")
     data.extend(id=uuid.uuid4())
     # Create the record in the database
     result = Notification.objects.create(data)
@@ -19,12 +24,18 @@ def post(data, current_user):
     return NotificationSchema.from_orm(result).dict()
 
 
-def delete(id, current_user):
+def delete(notification_id, current_user):
     """DELETE LOGIC"""
     try:
+        # Check if user is GlobalViewAdmin or has memberships
+        if not is_global_view_admin(current_user) and not get_org_memberships(
+            current_user
+        ):
+            raise HTTPException(status_code=403, detail="Unauthorized")
+
         # Validate that key_id is a valid UUID
-        uuid.UUID(id)
-        result = Notification.objects.get(id=id, userId=current_user)
+        uuid.UUID(notification_id)
+        result = Notification.objects.get(id=notification_id, userId=current_user)
 
         # Delete the Item
         result.delete()
@@ -52,11 +63,17 @@ def get_all():
         raise HTTPException(status_code=500, detail=str(error))
 
 
-def get_by_id(id, current_user):
+def get_by_id(notification_id, current_user):
     """GET by id"""
     try:
+        # Check if user is GlobalViewAdmin or has memberships
+        if not is_global_view_admin(current_user) and not get_org_memberships(
+            current_user
+        ):
+            raise HTTPException(status_code=403, detail="Unauthorized")
+
         # Find the item by its id
-        result = Notification.objects.get(id=id)
+        result = Notification.objects.get(id=notification_id)
 
         # Convert the result to Schema using from_orm
         return NotificationSchema.from_orm(result)
@@ -67,16 +84,17 @@ def get_by_id(id, current_user):
         raise HTTPException(status_code=500, detail=str(error))
 
 
-def get_508_banner(current_user):
-    """GET 508 banner."""
-    # TODO: Adding placeholder until we determine if we still need this.
-    # Remove logic if no longer needed or update to actual return object.
-    try:
-        # Get the 508 banner from the DB
-        result = ""
+# TODO: Adding placeholder until we determine if we still need this.
+# def get_508_banner(current_user):
+#     """GET 508 banner."""
 
-        # Format/Return Banner
-        return result
+#     # Remove logic if no longer needed or update to actual return object.
+#     try:
+#         # Get the 508 banner from the DB
+#         result = ""
 
-    except Exception as error:
-        raise HTTPException(status_code=500, detail=str(error))
+#         # Format/Return Banner
+#         return result
+
+#     except Exception as error:
+#         raise HTTPException(status_code=500, detail=str(error))
